got backend working
