+++ .env
@@ -1,0 +1,4 @@
+DB_HOST=localhost
+DB_USER=root
+DB_PASSWORD=admin
+DB_DATABASE=job_journey_db
