--- backend/.env
@@ -1,4 +1,0 @@
-DB_HOST=localhost
-DB_USER=root
-DB_PASSWORD=root
-DB_DATABASE=api
