+++ src/App.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { Routes, Route } from 'react-router-dom';
 import './App.scss';
 import Home from './components/Home/Home';
