+++ src/components/Application/Application.tsx
@@ -1,6 +1,3 @@
-import React from 'react';
-import styles from './Application.module.scss';
-
 const Application = (): JSX.Element => {
   return (
     <div>Application Page!</div>
