+++ src/components/Home/Home.tsx
@@ -1,6 +1,3 @@
-import React from 'react';
-import styles from './Home.module.scss';
-
 const Home = (): JSX.Element => {
   return (
     <div>Home Page!</div>
