+++ src/components/Resume/Resume.tsx
@@ -1,6 +1,3 @@
-import React from 'react';
-import styles from './Resume.module.scss';
-
 const Resume = (): JSX.Element => {
   return (
     <div>Resume Page!</div>
